Refactor actionTemplateBase.ts
