+++ _common/gameModel/libs/ClientScript/dashboard/dashboardFacade.ts
@@ -4,7 +4,7 @@
  * Put minimal logic in here.
  */
 
-import { MapChoiceActionTemplate, SimFlag } from '../game/common/actions/actionTemplateBase';
+import { SimFlag } from '../game/common/actions/actionTemplate/actionTemplateBase';
 import { InterventionRole } from '../game/common/actors/actor';
 import { getRoleLongTranslation, getRoleShortTranslation } from '../game/common/actors/actorLogic';
 import {
@@ -59,6 +59,7 @@ import {
 import { dashboardLogger } from '../tools/logger';
 import { MapEntityActivable } from '../game/common/simulationState/activableState';
 import { getActionTemplates } from '../game/mainSimulationLogic';
+import { MapChoiceActionTemplate } from '../game/common/actions/actionTemplate/actionTemplateBaseMap';
 
 // -------------------------------------------------------------------------------------------------
 // state part
