+++ _common/gameModel/libs/ClientScript/game/common/actions/actionTemplateDescriptor/descriptors/moveTemplate.ts
@@ -1,6 +1,6 @@
 import { TimeSliceDuration } from '../../../constants';
-import { MoveActorActionTemplate } from '../../actionTemplateBase';
 import { ITemplateDescriptor } from '../../actionTemplateDescriptor/templateDescriptor';
+import { MoveActorActionTemplate } from '../../actionTemplate/actionTemplateBaseActors';
 
 export interface MoveActorTemplateDescriptor extends ITemplateDescriptor {
   type: 'MoveActorTemplateDescriptor';
