+++ _common/gameModel/libs/ClientScript/game/loaders/actionTemplateFactory.ts
@@ -1,4 +1,4 @@
-import { ActionTemplateBase } from '../common/actions/actionTemplateBase';
+import { ActionTemplateBase } from '../common/actions/actionTemplate/actionTemplateBase';
 import { createMapChoiceActionTemplate as createMapChoiceActionTemplate } from '../common/actions/actionTemplateDescriptor/descriptors/mapChoiceTemplate';
 import { createFullyConfigurableTemplate } from '../common/actions/actionTemplateDescriptor/descriptors/fullyConfigurableTemplate';
 import { createMoveActorTemplate } from '../common/actions/actionTemplateDescriptor/descriptors/moveTemplate';
