+++ _common/gameModel/libs/ClientScript/game/loaders/resourceLoader.ts
@@ -1,6 +1,6 @@
 import { resourceLogger } from '../../tools/logger';
 import { parseObjectDescriptor } from '../../tools/WegasHelper';
-import { SimFlag } from '../common/actions/actionTemplateBase';
+import { SimFlag } from '../common/actions/actionTemplate/actionTemplateBase';
 import { InterventionRole } from '../common/actors/actor';
 import { ResourceContainerDefinitionId, TranslationKey } from '../common/baseTypes';
 import {
